refactor(api): rename mod api.rs -> batch.rs
