+++ habitat-api/src/batch.rs
